# api-contract/openapi.yaml
openapi: 3.1.0
info:
  title: MoringaLab Commerce API (`فروشگاه سبزینه`)
  description: Official OpenAPI 3.1 REST API contract between Go Backend Monolith and Next.js Storefront/Admin.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Development Local API Server
paths:
...
  /admin/access/search-users:
    get:
      summary: Search existing registered users to promote to admin
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Matching users list
...